requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.13
psycopg-pool==3.2.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
python-multipart==0.0.9
bcrypt==4.1.3
python-dotenv>=1.0.0
openpyxl==3.1.5
